requirement.txt
flask
openai
python-dotenv
requests
markdown
flask_cors
pymysql
chromadb
sentence-transformers
groq